Moved data files to data folder
