+++ src/gui/data_structures/sky_data.cpp
@@ -1,7 +1,5 @@
 #include "sky_data.hpp"
 
-#include "../logging.hpp"
-
 #include <GL/glew.h>
 #include <GLFW/glfw3.h>
 #include <glm/glm.hpp>
