+++ src/gui/data_structures/sky_data.hpp
@@ -1,5 +1,3 @@
-#include "../logging.hpp"
-
 #include <GL/glew.h>
 #include <GLFW/glfw3.h>
 #include <glm/glm.hpp>
