+++ src/gui/main_gui.cpp
@@ -9,8 +9,8 @@
 #include "controls.hpp"
 #include "render/renderer.hpp"
 #include "shader.hpp"
-#include "screen_data.hpp"
-#include "sky_data.hpp"
+#include "data_structures/screen_data.hpp"
+#include "data_structures/sky_data.hpp"
 #include "render/shadow_map.hpp"
 #include "render/sky.hpp"
 #include "gui_logging.hpp"
